+++ src/utils.cpp
@@ -1,7 +1,7 @@
 #include "../inc/utils.hpp"
 
 
-IntArray::IntArray(int length) : length(length) {
+IntArray::IntArray(uint64_t length) : length(length) {
     data = new int[length];
     if (!data) {
         throw std::runtime_error("IntArray: Memory allocation failed for array of length " + std::to_string(length));
@@ -15,14 +15,14 @@ IntArray::~IntArray() {
 
 namespace ArrayUtils {
     
-    void fill_array_random(IntArray& array, int max_value) {
-        for (int i = 0; i < array.length; i++) {
+    void fill_array_random(IntArray& array, uint64_t max_value) {
+        for (uint64_t i = 0; i < array.length; i++) {
             array.data[i] = std::rand() % max_value;
         }
     }
 
     void print_arr(const IntArray& array) {
-        for (int i = 0; i < array.length; i++) {
+        for (uint64_t i = 0; i < array.length; i++) {
             std::cout << array.data[i] << " ";
         }
         std::cout << std::endl << std::endl;
